artcase/requirements.txt
Django==1.8.2
django-ckeditor==4.4.8
django-date-extensions==1.0.0
django-dbbackup==2.0.4
django-extensions==1.5.5
django-widget-tweaks==1.3
ecdsa==0.13
paramiko==1.15.2
Pillow==2.8.1
pycrypto==2.6.1
pysftp==0.2.8
six==1.9.0
slugify==0.0.1
Unidecode==0.4.17
Unipath==1.1
